// components/Questionnaire/types/types.ts
export interface FormQuestionnaire {
    firstName: string;
    birthday: string;
    sex: string;
    height: number | null;
    weight: number | null;
    bodyBuild: string;
    eyeColor: string;
    aboutMe: string;
    interests: Array<string>;
    character: string;
    familyStatus: string;
    orientation: string;
    looking: Array<string>;
    qualities: Array<string>;
    partnerDesc: string;
}
  